Extract an action handler for FIAM clicks
